PydubLoader: Add helper class for SoundFileLoader to avoid repetition
There needs to be some class sitting between SoundFileLoader and the
concrete children since
 - each child shares a lot of code with siblings
 - each child needs the `pydub` external library, which may not exist
   on all systems

See #122 for more context.
